+++ app/Http/Controllers/ListeningControllers/ListeningExamController.php
@@ -2,10 +2,12 @@
 
 namespace App\Http\Controllers\ListeningControllers;
 
+use App\Exam;
 use App\GroupType;
 use App\Http\Controllers\Controller;
 use App\Listening\Audio;
 use App\Listening\ListeningExam;
+use App\Logging;
 use App\Reservation;
 use Illuminate\Http\Request;
 
@@ -21,7 +23,7 @@ class ListeningExamController extends Controller
                 'actions' => '',
             ];
         });
-        $exams=json_encode($exams);
+        $exams = json_encode($exams);
         return view('listening.exams.index', compact('exams'));
     }
 
@@ -57,12 +59,16 @@ class ListeningExamController extends Controller
         $count = ListeningExam::where('reservation_id', $res)->count();
 
         if ($count == 0) {
-            ListeningExam::create([
+            $exam = ListeningExam::create([
                 'reservation_id' => $res,
 //                'group_type_id'=>$type,
             ]);
+            $message = " add new listening exam with id {" . $exam->id . "} for reservation with id {" . $res . "}";
+            Logging::logProfessor(auth()->user(), $message);
             return redirect()->to(route('listening.index'));
         } else {
+            $message = " have made exam to this reservation with id {" . $res . "}";
+            Logging::logProfessor(auth()->user(), $message);
             return redirect()->back()->with('error', 'You have made exam to this reservation');
         }
 
@@ -78,7 +84,7 @@ class ListeningExamController extends Controller
     {
 
         $audios = Audio::getAudios($exam->audios()->get());
-        $audios=json_encode($audios);
+        $audios = json_encode($audios);
 
         return view('listening.exams.show', compact('audios', 'exam'));
     }
@@ -110,15 +116,22 @@ class ListeningExamController extends Controller
 //        $type=intval($request['type']);
 //        $count =ListeningExam::where('reservation_id',$res)->where('group_type_id',$type)->count();
         $count = ListeningExam::where('reservation_id', $res)->count();
-
-        if ($count == 0) {
-            $exam->update([
-                'reservation_id' => intval($request['reservation']),
+        if (!Exam::checkExamIsRunning($exam)) {
+            if ($count == 0) {
+                $message = " update listening exam with id {" . $exam->id . "} and old reservation is {" . $exam->reservation->id . "} and new reservation id is{" . $res . "}";
+                Logging::logProfessor(auth()->user(), $message);
+                $exam->update([
+                    'reservation_id' => intval($request['reservation']),
 //                'group_type_id'=>intval($request['type']),
-            ]);
-            return redirect()->to(route('listening.index'));
+                ]);
+                return redirect()->to(route('listening.index'));
+            } else {
+                $message = " have made exam to this reservation with id {" . $res . "}";
+                Logging::logProfessor(auth()->user(), $message);
+                return redirect()->back()->with('error', 'You have made exam to this reservation ');
+            }
         } else {
-            return redirect()->back()->with('error', 'You have made exam to this reservation ');
+            return redirect()->back()->with('error', 'You can\'t edit this exam because it is running ');
         }
     }
 
@@ -131,9 +144,13 @@ class ListeningExamController extends Controller
      */
     public function destroy(ListeningExam $exam)
     {
-        $exam->audios()->detach($exam->audios);
-        $exam->delete();
+        if (!Exam::checkExamIsRunning($exam)) {
+            $message = " delete listening exam with id {" . $exam->id . "} and  reservation id is {" . $exam->reservation->id . "}";
+            Logging::logProfessor(auth()->user(), $message);
+            $exam->audios()->detach($exam->audios);
+            $exam->delete();
 //        return redirect()->action('ListeningExamController@index');
+        }
     }
